// src/app/visitantes/models/interfaces/visitantes.interface.ts
// Generated by https://quicktype.io

export interface IVisitanteRequest {
    nombre: string;
    apellidoPaterno: string;
    apellidoMaterno: string;
    edad: number;
    email: string;
    username: string;
    telefono: string;
    direccion: string;
}

export interface IVisitanteResponse {
    nombre: string;
    apellidoPaterno: string;
    apellidoMaterno: string;
    edad: number;
    email: string;
    username: string;
    telefono: string;
    direccion: string;
}